requirements.txt
appdirs==1.4.3
astroid==2.2.5
attrs==19.1.0
black==19.3b0
Click==7.0
entrypoints==0.3
flake8==3.7.7
isort==4.3.18
lazy-object-proxy==1.4.0
mccabe==0.6.1
numpy==1.16.3
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.3.1
PyYAML==5.1.1
six==1.12.0
toml==0.10.0
typed-ast==1.3.5
wrapt==1.11.1
